updated CSS for avatar selection
