+++ src/app/(pages)/profile/page.module.css
@@ -141,13 +141,17 @@
 }
 
 .active {
-  /* border-color: red; */
-  border: 3px red solid;
+  width: 100px; /* Adjust the width dynamically */
+  height: 100px;
+  border-radius: 54px;
+  background-size: cover;
+  margin: 30px 10px;
+  border: 5px solid green;
+  transition: border-color 0.3s ease;
 }
 
 .popPic:hover {
-  border-color: green;
-  border: 3px solid;
+  opacity: 0.5;
 }
 
 .popBottom {
